# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Trend musix
host: localhost:10010
basePath: /
...
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /track/get-track:
    x-swagger-router-controller: track
    get:
      description: Get track's data
      operationId: getTrack
      parameters:
        - name: id
          in: query
          description: Track's id
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/TrackResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /track/get-home-track:
    x-swagger-router-controller: track
    get:
      description: Get track's data for home page
      operationId: getHomeTrack
      parameters:
        - name: key
          in: query
          description: Token
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/TrackResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /user/favorite:
    x-swagger-router-controller: user
    get:
      description: Get list favorite song
      operationId: getFavoriteSong
      parameters:
        - name: token
          in: query
          description: Token
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ListTrackInfoResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Response"
  /track/search:
    x-swagger-router-controller: track
    get:
      description: Search track
      operationId: searchTrack
      parameters:
        - name: keyword
          in: query
          description: keyword
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ListTrackInfoResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Response"
...
  /artist/search:
    x-swagger-router-controller: track
    get:
      description: Search artist
      operationId: searchArtist
      parameters:
        - name: keyword
          in: query
          description: keyword
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ListArtistResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Response"
...
definitions:
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  TrackInfoResponse:
    properties:
      id:
        type: string
      title:
        type: string
      artist:
        type: string
      artist_imageurl:
        type: string
      genre:
        type: string
      genre_imageurl:
        type: string
      imageurl:
        type: string
      url:
        type: string
      like:
        type: number
      listen:
        type: number
  TrackFeatureResponse:
    properties:
      speechiness:
        type: number
      acousticness:
        type: number
      instrumentalness:
        type: number
      liveness:
        type: number
      valence:
        type: number
      duration_ms:
        type: number
      tempo:
        type: number
      time_signature:
        type: number
      mode:
        type: number
      key:
        type: number
      loudness:
        type: number
      danceability:
        type: number
      energy:
        type: number
  TrackResponse:
    properties:
      info:
        type: object
        $ref: "#/definitions/TrackInfoResponse"
      lyric:
        type: string
      feature:
        type: object
        $ref: "#/definitions/TrackFeatureResponse"
      hit:
        type: boolean
...
  ListTrackInfoResponse:
    properties:
      status:
        type: integer
      listTrackInfo:
        type: array
        items:
          $ref: "#/definitions/TrackInfoResponse"
  ListArtistResponse:
    properties:
      status:
        type: integer
      listArtistResponse:
        type: array
        items:
          $ref: "#/definitions/ArtistResponse"
  ArtistResponse:
    properties:
      id:
        type: string
      name:
        type: string
      imageurl:
        type: string
  Response:
    properties:
      status:
        type: integer
      message:
        type: string